requirements.txt
streamlit>=1.31.0
phidata>=2.3.0
ollama>=0.1.6